requirements.txt
alembic==1.13.1
bidict==0.23.1
blinker==1.8.2
click==8.1.7
dominate==2.9.1
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Cors==4.0.1
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
simple-websocket==1.0.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
visitor==0.1.3
Werkzeug==3.0.3
wsproto==1.2.0

APScheduler~=3.10.4